requirements.txt
requests
slack_sdk
python-dotenv
schedule
pymupdf
